machine-learning-client/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
astroid==3.1.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
black==24.3.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
datasets==2.18.0
decorator==5.1.1
dill==0.3.8
dnspython==2.6.1
exceptiongroup==1.2.0
filelock==3.13.4
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.2.0
huggingface-hub==0.22.2
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.0
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.42.0
MarkupSafe==2.1.5
mccabe==0.7.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pooch==1.8.1
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycparser==2.22
pyee==11.1.0
pylint==3.1.0
pymongo==4.6.3
pytest==8.1.1
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ffmpeg==2.0.10
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
sentencepiece==0.2.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
threadpoolctl==3.4.0
tokenizers==0.15.2
tomli==2.0.1
tomlkit==0.12.4
torch==2.2.2
torchaudio==2.2.2
tqdm==4.66.2
transformers==4.39.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
xxhash==3.4.1
yarl==1.9.4
